Séparer les modèles de SV en plusieurs fichiers
Afin d'améliorer la lisibilité et la maintenance.
