+++ sv/models/__init__.py
@@ -1,0 +1,5 @@
+from .common import *  # noqa F403
+from .fiches_detection import *  # noqa F403
+from .fiches_zone_delimitee import *  # noqa F403
+from .lieux import *  # noqa F403
+from .prelevements import *  # noqa F403
